requirements.txt
numpy
pandas
matplotlib
seaborn
sklearn.utils
flask